Changed order of preference for solvers: Gurobi > bpmpd > osqp > qpOASES
Notice that while this change seems trivial, it actually causes tests to pass.
Since bpmpd interface is quite brittle, it was the case that using it as a third option
in tests caused some of them to fail. This means the order of execution of tests
has an influence on the solver, which is a bad sign.