Adds a unit test to test that the examples run.
Removes static global KDLEnvPtr
This is a workaround to avoid the ClassUnloader exception. I am not sure why this was causing a problem, but it should be explored further. When the pointer was defined inside main, the problem went away.

Test both json and cpp methods

Bug fix in car seat demo

Bug fix to basic cartesian example

Fix Cmake for examples

Add pick and place plan test